+++ Dockerfile
@@ -29,7 +29,7 @@ RUN ARCHIVE=unknown && \
 
 
 WORKDIR /app
-COPY . .
+COPY app/ .
 
 EXPOSE 5000
 CMD ["gunicorn", "-b", "0.0.0.0:5000", "app:app"]
