+++ app/lldap_graphql/const.py
