+++ app/templates/base.html
@@ -1,0 +1,20 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="utf-8" />
+  <link rel="icon" type="image/x-icon" href="{{ url_for('static', filename='favicon.ico') }}">
+  <title>{% block title %}lldap-request{% endblock %}</title>
+  {% block extra_head %}{% endblock %}
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <script src="https://cdn.tailwindcss.com"></script>
+</head>
+<body class="{% block body_class %}bg-gray-100 text-gray-900 font-sans{% endblock %}">
+  {% block content %}{% endblock %}
+
+  {% block footer %}
+  <footer class="text-center text-xs text-gray-400 mt-8 mb-2">
+    lldap-request {{ version }}
+  </footer>
+  {% endblock %}
+</body>
+</html>
