+++ pyproject.toml
@@ -206,6 +206,7 @@ ignore = [
     "COM812",
     "COM819",
     "ISC001",
+    "INP001",
 
     # Disabled because ruff does not understand type of __all__ generated by a function
     "PLE0605"
