+++ requirements.txt
@@ -1,4 +1,5 @@
 flask
+flask-wtf
 gunicorn
 requests
 mypy
