--- templates/index.html
@@ -1,45 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8" />
-  <title>Request an Account</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body class="bg-gray-100 text-gray-900 font-sans">
-
-  <div class="max-w-xl mx-auto mt-10 p-6 bg-white rounded-2xl shadow-md">
-    <h1 class="text-2xl font-bold mb-4 text-center">Request an Account</h1>
-    <form method="POST" action="/submit" class="space-y-4">
-      {% set labels = {
-        'username':    'Username',
-        'email':       'Email',
-        'displayname': 'Display name',
-        'firstname':   'First name',
-        'lastname':    'Last name'
-      } %}
-
-      {% for field in ['username', 'email', 'displayname', 'firstname', 'lastname'] %}
-        <div>
-          <label class="block font-medium capitalize mb-1" for="{{ field }}">{{
-            labels.get(field, field.replace('_', ' '))
-          }}</label>
-          <input
-            class="w-full border border-gray-300 rounded-lg px-3 py-2"
-            type="text"
-            id="{{ field }}"
-            name="{{ field }}"
-            {% if field in ['username', 'email'] %} required {% endif %}
-          >
-        </div>
-      {% endfor %}
-      <div class="text-center">
-        <button class="bg-blue-600 text-white px-4 py-2 rounded-lg hover:bg-blue-700 transition">Submit Request</button>
-      </div>
-    </form>
-  </div>
-  <footer class="text-center text-xs text-gray-400 mt-8 mb-2">
-    lldap-request {{ version }}
-  </footer>
-</body>
-</html>
