--- templates/submitted.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8" />
-  <title>Request Submitted</title>
-  <meta http-equiv="refresh" content="{{ delay }};url={{ redirect_url }}">
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body class="bg-gray-100 text-gray-900 font-sans flex items-center justify-center min-h-screen">
-
-  <div class="bg-white rounded-2xl shadow-md max-w-md w-full p-6 text-center">
-    <h1 class="text-2xl font-bold text-green-600 mb-4">{{ message }}</h1>
-    <p class="text-sm text-gray-600">
-      You will be redirected shortly
-      <br />
-      If not, <a href="{{ redirect_url }}" class="text-blue-600 underline">click here</a>
-    </p>
-
-    <footer class="text-center text-xs text-gray-400 mt-8">
-      lldap-request {{ version }}
-    </footer>
-  </div>
-
-</body>
-</html>
